merged wip-brenier branch to trunc -r 553:655 openvibe-plugins-acquisition
merged wip-brenier branch to trunc -r 547:655 openvibe-plugins-classification
merged wip-brenier branch to trunc -r 495:655 openvibe-plugins-feature-extraction
merged wip-brenier branch to trunc -r 558:655 openvibe-plugins-file-io
merged wip-yrenard branch to trunc -r 531:655 openvibe-plugins-samples
merged wip-brenier branch to trunc -r 552:655 openvibe-plugins-signal-processing
merged wip-brenier branch to trunc -r 499:655 openvibe-plugins-simple-visualisation
merged wip-brenier branch to trunc -r 497:655 openvibe-plugins-stimulation

openvibe :
 + added log and scenario managers access to player context
 + added log level benchmark
 + added player manager
 + added IPluginModuleContext object for the three plugin module global functions, so it can register types etc
 + added use hint on box proto
 * reorganised files organisation
 * modified .h definitions and class id names
 * changed clock frequency to be 32:32 fixed point hz

openvibe-kernel :
 + added correct time support for clock messages and frequency management
 + added plugin module context parameter to three functions of the plugin module
 + added player manager
 + added use hint on box proto
 + added log and scenario manager access to player context
 + added binding functions to ease creation of bridges
 + added bridge for log manager in the player context
 + added bridge for scenario manager in the player context
 + added log level benchmark
 + added benchmark chronometers, gives CPU usage percentage and time consuming in each box
 * reorganised files organisation
 * reorganised source tree
 * modified .h definitions and class id names
 * changed clock frequency to be 32:32 fixed point hz
 * enhanced plugin loader message trace (with error message printed on module load error)
 * changed some log levels

openvibe-toolkit :
 + added training template, signal trial and signal trial set
 + defined some stimulations identifiers
 * changed clock frequency to be 32:32 fixed point hz
 * enhanced box algorithm base class for output writing

openvibe-applications-designer :
 + added filename type setting
 * replaced player insciation / deletion with player manager functions
 * enhanced setting window behavior
 * enhanced log
 * modified some log levels
 * sorted box algorithm categories / names in alphabetical order

openvibe-modules-automaton :
 + added switch event node
 + WaitTime nodes now support random duration

openvibe-plugins-acquisition :
 * added frequency to generic network acquisition box
 * plugins now use OpenViBE logs

openvibe-plugins-classification :
 + added LDA Classifier plugin
 + added documentation for the classifier plugin
 * plugins now use OpenViBE logs

openvibe-plugins-feature-extraction :
 + added documentation for the feature aggregator plugin
 * fixed bug in Feature Aggregator in the buffer copy
 * fixed bug in Feature Aggregator in the label generation
 * plugins now use OpenViBE logs

openvibe-plugins-file-io :
 + added offline mode in GDF file reader
 + added BCI Competition III - Dataset IIIb reader (text format)
 + added documentation for the BCI competition IIIb reader
 * GDF writer now saves events (in GDF mode 1)
 * GDF writer moved chunk processing from processInput to process method
 * GDF writer fixed error with DIGITAL_MIN initialization value
 * GDF writer filename parameter is now of type Filename
 * GDF writer don't save trailing \0 character from channel names
 * GDF reader now correctly sets its clock frequency
 * GDF reader fixed a bug in sampling rate computation
 * GDF reader now supports event mode 3
 * GDF reader float32 and float64 data type samples aren't scaled anymore
 * GDF reader filename parameter is now of type Filename
 * Generic reader now checks chunks starting date
 * Generic reader fixed type size problem in getClockFrequency
 * plugins now use OpenViBE logs

openvibe-plugins-samples :
 + added log sample box (a box that logs stuffs on several events/stages, just so you know)
 + added test training algorithm (work in progress)
 * corrected bug on producer that caused the box to crash at uninitialization phase
 * corrected several box frequencies
 * changed clock frequency to be 32:32 fixed point hz
 * plugins now use OpenViBE logs

openvibe-plugins-simple-visualisation:
 + added a first version of a time-frequency analysis
 * simple signal display : decreased channel display widget's minimum height
 * simple signal display : left ruler is now disabled by default
 * simple signal display : changed input to Streamed Matrix
 * simple signal display : bottom ruler is now always displayed at the bottom of the window
 * simple signal display : bottom ruler (temporal) now displays dates in a human readable format :D
 * simple signal display : bottom ruler is now global for all the channels (displayed at the bottom)
 * simple signal display : buttons are now desactivated until the first buffer arrives
 * simple signal display : created a left ruler object
 * simple signal display : improved the left ruler labels display
 * simple signal display : fixed the bottom ruler alignement bug
 * simple signal display : fixed division by 0 error in sampling rate computation.
 * frequency spectrum display : improved display method (now uses rgbbuffers)
 + frequency spectrum display : the minmax attenuation coefficient is now configurable
 * Graz visualization : bar now grows according to Amplitude input
 * Graz visualization : now displays arrows and bars using png images
 * Graz visualization : changed left and right arrow images
 * Graz visualization : modified background and foreground color (green on black)
 * Graz visualization : changed bar scaling computation (now based on the maximum amplitude value per trial)
 * corrected two bugs that prevented best size mode from working properly
 * moved and renamed SignalDislayDatabase to BufferDatabase
 * moved ovpCSignalDislayBottomRuler to ovpCBottomTimeRuler
 * now possible to get min/max values for the currently displayed buffers, the last buffer, by channel, or for all channels.
 * plugins now use OpenViBE logs

openvibe-plugins-stimulation:
 + added documentation for the keyboard stimulator plugin and the XML Stimulation scenario player plugin
 + added an example stimulation scenario based on the Graz BCI
 * clock based plugins now correctly implements getClockFrequency method
 * keyboard stimulator : now sends empty stimulation buffers when there are no keyboard input
 * keyboard stimulator : improved error handling
 * keyboard stimulator : now based on a visible window which needs focus
 * plugins now use OpenViBE logs


git-svn-id: svn://scm.gforge.inria.fr/svn/openvibe@656 c330d7e9-fc0c-0410-a5b3-fd85c6f5aa8f

+++ openvibe/trunc/src/openvibe/plugins/ovIBoxAlgorithm.h
@@ -59,15 +59,17 @@ namespace OpenViBE
 			 * \note Default implementation returns 0
 			 *
 			 * This function is used for algorithms that are triggered on
-			 * clock signals. The frequency is given in Hz, thus returning
-			 * 1 will make the algorithm to be called every second,
-			 * returning 100 will make the algorithm being called 100 times
-			 * per second and so on...
+			 * clock signals. The frequency is given in Hz, with 32:32 fixed
+			 * point representation thus returning \c (1<<32) will make the
+			 * algorithm to be called every second, returning \c (100<<32)
+			 * will make the algorithm being called 100 times a second, 
+			 * returning \c (1<<31) will make the algorithm be called once
+			 * every two seconds and so on...
 			 *
 			 * \note Returning 0 means the algorithm should not be
 			 *       clock activated.
 			 */
-			virtual OpenViBE::uint32 getClockFrequency(
+			virtual OpenViBE::uint64 getClockFrequency(
 				OpenViBE::Kernel::IBoxAlgorithmContext& rBoxAlgorithmContext) { return 0; }
 
 			//@}
